cleaned up main.cc - moved system monitoring out
