+++ CMakeLists.txt
@@ -251,6 +251,7 @@ list(APPEND trunk_recorder_sources
   trunk-recorder/csv_helper.cc
   trunk-recorder/config.cc
   trunk-recorder/setup_systems.cc
+  trunk-recorder/monitor_systems.cc
   trunk-recorder/talkgroup.cc
   trunk-recorder/talkgroups.cc
   trunk-recorder/unit_tag.cc
